Fix lastapp action going to homescreen when no default launcher is set
when no launcher is set as default in:
Settings - Apps and notifications - Advanced - Default apps - Home
(can happen for example when the user starts trying more launchers
without choosing the default one after pressing the home button, and
this state is kept also after uninstalling all of them).

if the user was on home screen before running the current app,
then triggers a last app action (e.g. from recents double tap
if set in du navbar settings), the home screen is shown instead
of the previously running app.

This happens because pm.resolveActivity gives "android" or the resolve
activity as result in this case.
We can use instead pm.getHomeActivities (a hidden api that basically
triggers the public queryIntentActivities) to correctly skip all
available launchers when comparing the list of last running apps.

Tecnically this method could be slower than resolveActivity but it's more
reliable and anyway i can't see any delay.

Change-Id: I0569b5be4c1cc4b20fcc0499273b5d7b9395402f
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>